split audits specific to rawmatrixfile from analysisfile
